Adding packages/classes: - MainPage - IndexPage Modifying SignIn - use PageObjects
