+++ src/test/java/SignIn.java
@@ -4,52 +4,32 @@ import org.openqa.selenium.support.ui.ExpectedCondition;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
 import org.testng.annotations.Test;
+import pages.IndexPage;
 
 public class SignIn {
     WebDriver driver = new ChromeDriver();
 
-    public boolean waitForJStoLoad() {
-        WebDriverWait wait = new WebDriverWait(driver, 30);
 
-        try {
-            Thread.sleep(100);  //sleep 100ms to be sure there is no delay in browser befor kicking of jquery or js
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
 
-        // wait for jQuery to load
-        ExpectedCondition<Boolean> jQueryLoad = new ExpectedCondition<Boolean>() {
-            public Boolean apply(WebDriver driver) {
-                try {
-                    return ((Long) ((JavascriptExecutor) driver).executeScript("return jQuery.active") == 0);
-                } catch (Exception e) {
-                    return true;
-                }
-            }
-        };
-
-        // wait for Javascript to load
-        ExpectedCondition<Boolean> jsLoad = new ExpectedCondition<Boolean>() {
-            public Boolean apply(WebDriver driver) {
-                return ((JavascriptExecutor) driver).executeScript("return document.readyState")
-                        .toString().equals("complete");
-            }
-        };
-        return wait.until(jQueryLoad) && wait.until(jsLoad);
-    }
 
     public void login (String userName, String password) {
+
         driver.findElement(By.cssSelector("[name='username']")).sendKeys(userName);
         driver.findElement(By.cssSelector("[name='password']")).sendKeys(password);
         driver.findElement(By.cssSelector("div .login .button")).click();
     }
 
-    @Test(priority = 1)
+    @Test
     public void shouldLogin() {
-        driver.get("http://parabank.parasoft.com");
-        waitForJStoLoad();
-        login("test", "test");
-        waitForJStoLoad();
+
+        IndexPage indexPage = new IndexPage(driver);
+
+        indexPage.openParaBank();
+        indexPage.waitForJStoLoad();
+        indexPage.fillUserName("test");
+        indexPage.fillPassword("password");
+        indexPage.clickLogin();
+        indexPage.waitForJStoLoad();
         Assert.assertTrue(driver.findElement(By.cssSelector("a[href$='logout.htm']")).isDisplayed());
     }
